Fix signup flow and update branding

+++ system/functions/functions.php
@@ -429,7 +429,7 @@ function validip($ip) {
 //Формирование секретного кода
 function mksecret($length = 32) {
 $set = array("a","A","b","B","c","C","d","D","e","E","f","F","g","G","h","H","i","I","j","J","k","K","l","L","m","M","n","N","o","O","p","P","q","Q","r","R","s","S","t","T","u","U","v","V","w","W","x","X","y","Y","z","Z","1","2","3","4","5","6","7","8","9");
-	$str;
+	$str = '';
 	for($i = 1; $i <= $length; $i++)
 	{
 		$ch = rand(0, count($set)-1);
@@ -486,7 +486,9 @@ function logout_cookie() {
 	setcookie(COOKIE_ID, "", 0x7fffffff, "/" , $domain , false , true); 
 	setcookie(COOKIE_PASSWORD, "", 0x7fffffff, "/" , $domain , false , true); 
 	//Удаляем memcached файл
-	$memcache->delete('user_'.$USER['id']);
+	if($USER && isset($USER['id'])) {
+		$memcache->delete('user_'.$USER['id']);
+	}
 }
 
 //Проверка авторизации
